Refactor image and video gallery components to improve structure and functionality. Removed redundant code and integrated new ImageCard and VideoCard components for better modularity. Enhanced error handling and user experience in image and video displays. Updated documentation to reflect changes and ensure clarity in component usage.
