+++ app/tools/video-generator/components/video-gallery.tsx
@@ -4,16 +4,12 @@ import { useState } from 'react';
 import { Button } from '@/components/ui/button';
 import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card';
 import { 
-  Download, 
-  Copy, 
   Trash2, 
-  Play,
-  Clock,
-  Settings,
-  X,
-  Video
+  Video,
 } from 'lucide-react';
 import type { GeneratedVideo } from '../hooks/use-video-generator';
+import { VideoPreviewModal } from './video-preview-modal';
+import { VideoCard } from './video-card';
 
 interface VideoGalleryProps {
   videos: GeneratedVideo[];
@@ -35,185 +31,8 @@ export function VideoGallery({
   const [selectedVideo, setSelectedVideo] = useState<GeneratedVideo | null>(null);
   const [videoErrors, setVideoErrors] = useState<Set<string>>(new Set());
 
-  // AICODE-NOTE: Handle video load errors
   const handleVideoError = (videoId: string) => {
     setVideoErrors(prev => new Set(prev).add(videoId));
-  };
-
-  // AICODE-NOTE: Format timestamp for display
-  const formatTimestamp = (timestamp: number): string => {
-    const date = new Date(timestamp);
-    const now = new Date();
-    const diffMs = now.getTime() - date.getTime();
-    const diffMins = Math.floor(diffMs / 60000);
-    
-    if (diffMins < 1) return 'Just now';
-    if (diffMins < 60) return `${diffMins}m ago`;
-    if (diffMins < 1440) return `${Math.floor(diffMins / 60)}h ago`;
-    
-    return date.toLocaleDateString();
-  };
-
-  // AICODE-NOTE: Format duration for display
-  const formatDuration = (seconds: number): string => {
-    if (seconds < 60) return `${seconds}s`;
-    const mins = Math.floor(seconds / 60);
-    const secs = seconds % 60;
-    return `${mins}:${secs.toString().padStart(2, '0')}`;
-  };
-
-  // AICODE-NOTE: Video preview modal
-  const VideoPreviewModal = ({ video }: { video: GeneratedVideo }) => (
-    <div className="fixed inset-0 bg-black/90 flex items-center justify-center z-50 p-4">
-      <div className="relative max-w-5xl max-h-full">
-        <Button
-          variant="outline"
-          size="sm"
-          className="absolute top-2 right-2 z-10 bg-gray-900/90 hover:bg-gray-800 text-white border-gray-600 size-8 p-0"
-          onClick={() => setSelectedVideo(null)}
-          title="Close"
-        >
-          <X className="size-3" />
-        </Button>
-        
-        <video
-          src={video.url}
-          controls
-          autoPlay
-          className="max-w-full max-h-full rounded-lg"
-          onError={() => handleVideoError(video.id)}
-        />
-        
-        <div className="absolute bottom-0 inset-x-0 bg-black/70 text-white p-4 rounded-b-lg">
-          <p className="text-sm font-medium line-clamp-2">{video.prompt}</p>
-          <div className="flex justify-between items-center mt-2 text-xs text-gray-300">
-            <span>{formatTimestamp(video.timestamp)}</span>
-            <span>{formatDuration(video.settings.duration)} • {video.settings.frameRate}fps</span>
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-
-  // AICODE-NOTE: Individual video card component
-  const VideoCard = ({ video, isCurrent = false }: { video: GeneratedVideo; isCurrent?: boolean }) => {
-    const hasError = videoErrors.has(video.id);
-    
-    return (
-      <Card className={`group relative overflow-hidden ${isCurrent ? 'ring-2 ring-blue-500' : ''}`}>
-        <div className="aspect-video relative bg-gray-900">
-          {hasError ? (
-            <div className="size-full bg-gray-100 flex items-center justify-center">
-              <div className="text-center text-gray-500">
-                <Video className="size-6 mx-auto mb-2" />
-                <p className="text-sm">Failed to load</p>
-              </div>
-            </div>
-          ) : (
-            <video
-              src={video.url}
-              className="size-full object-cover cursor-pointer"
-              muted
-              onMouseEnter={(e) => {
-                const video = e.currentTarget;
-                video.play().catch(() => {
-                  // Ignore autoplay errors
-                });
-              }}
-              onMouseLeave={(e) => {
-                const video = e.currentTarget;
-                video.pause();
-                video.currentTime = 0;
-              }}
-              onClick={() => setSelectedVideo(video)}
-              onError={() => handleVideoError(video.id)}
-            />
-          )}
-          
-          {/* Play button overlay */}
-          <div className="absolute inset-0 bg-black/0 group-hover:bg-black/30 transition-all duration-200 flex items-center justify-center">
-            <div className="opacity-0 group-hover:opacity-100 transition-opacity">
-              <Button
-                size="lg"
-                variant="outline"
-                className="bg-gray-900/90 hover:bg-gray-800 text-white border-gray-600 rounded-full size-14 p-0"
-                onClick={() => setSelectedVideo(video)}
-                title="Play video"
-              >
-                <Play className="size-5" />
-              </Button>
-            </div>
-          </div>
-          
-          {/* Action buttons */}
-          <div className="absolute top-2 right-2 opacity-0 group-hover:opacity-100 transition-opacity flex gap-1">
-            <Button
-              size="sm"
-              variant="outline"
-              className="bg-gray-900/80 hover:bg-gray-800 text-white border-gray-600 size-8 p-0"
-              onClick={() => onDownloadVideo(video)}
-              title="Download video"
-            >
-              <Download className="size-3" />
-            </Button>
-            <Button
-              size="sm"
-              variant="outline"
-              className="bg-gray-900/80 hover:bg-gray-800 text-white border-gray-600 size-8 p-0"
-              onClick={() => onCopyVideoUrl(video)}
-              title="Copy video URL"
-            >
-              <Copy className="size-3" />
-            </Button>
-            <Button
-              size="sm"
-              variant="outline"
-              className="bg-red-600/80 hover:bg-red-700 text-white border-red-500 size-8 p-0"
-              onClick={() => onDeleteVideo(video.id)}
-              title="Delete video"
-            >
-              <Trash2 className="size-3" />
-            </Button>
-          </div>
-          
-          {/* Duration badge */}
-          <div className="absolute bottom-2 right-2 bg-black/70 text-white text-xs px-2 py-1 rounded">
-            {formatDuration(video.settings.duration)}
-          </div>
-          
-          {isCurrent && (
-            <div className="absolute top-2 left-2">
-              <span className="bg-blue-500 text-white text-xs px-2 py-1 rounded-full">
-                Latest
-              </span>
-            </div>
-          )}
-        </div>
-        
-        <CardContent className="p-3">
-          <p className="text-sm font-medium line-clamp-2 mb-2">
-            {video.prompt}
-          </p>
-          
-          <div className="flex items-center justify-between text-xs text-gray-500">
-            <div className="flex items-center">
-              <Clock className="size-3 mr-1" />
-              {formatTimestamp(video.timestamp)}
-            </div>
-            
-            <div className="flex items-center">
-              <Settings className="size-3 mr-1" />
-              {video.settings.model}
-            </div>
-          </div>
-          
-          <div className="flex items-center justify-between text-xs text-gray-500 mt-1">
-            <span>{video.settings.resolution}</span>
-            <span>{video.settings.frameRate}fps</span>
-          </div>
-        </CardContent>
-      </Card>
-    );
   };
 
   if (videos.length === 0 && !currentGeneration) {
@@ -261,6 +80,12 @@ export function VideoGallery({
               <VideoCard 
                 video={currentGeneration} 
                 isCurrent={true}
+                onCopyVideoUrl={onCopyVideoUrl}
+                onDownloadVideo={onDownloadVideo}
+                onDeleteVideo={onDeleteVideo}
+                setSelectedVideo={setSelectedVideo}
+                handleVideoError={handleVideoError}
+                videoErrors={videoErrors}
               />
             )}
             
@@ -271,6 +96,12 @@ export function VideoGallery({
                 <VideoCard 
                   key={video.id} 
                   video={video}
+                  onCopyVideoUrl={onCopyVideoUrl}
+                  onDownloadVideo={onDownloadVideo}
+                  onDeleteVideo={onDeleteVideo}
+                  setSelectedVideo={setSelectedVideo}
+                  handleVideoError={handleVideoError}
+                  videoErrors={videoErrors}
                 />
               ))
             }
@@ -280,8 +111,15 @@ export function VideoGallery({
 
       {/* Video Preview Modal */}
       {selectedVideo && (
-        <VideoPreviewModal video={selectedVideo} />
-      )}
+        <VideoPreviewModal 
+          video={selectedVideo} 
+          setSelectedVideo={setSelectedVideo} 
+          handleVideoError={handleVideoError} 
+        />
+      )}  
     </>
   );
 } 
+
+
+
