+++ hooks/use-video-effects.ts
@@ -1,5 +1,6 @@
 import { useEffect, useRef } from 'react';
 import type { UseChatHelpers } from '@ai-sdk/react';
+import { saveMediaToChat } from '@/lib/ai/chat/media';
 
 interface UseVideoEffectsProps {
   videoUrl?: string;
@@ -24,125 +25,6 @@ const generateUUID = (): string => {
   });
 };
 
-// AICODE-NOTE: Function to check if video already exists in chat to prevent duplicates
-const checkVideoExistsInChat = (
-  setMessages: UseChatHelpers['setMessages'],
-  videoUrl: string,
-): boolean => {
-  let videoExists = false;
-
-  setMessages((prevMessages) => {
-    // Check if any message already contains this video URL
-    videoExists = prevMessages.some((message) =>
-      message.experimental_attachments?.some(
-        (attachment) => attachment.url === videoUrl,
-      ),
-    );
-    return prevMessages; // Don't modify messages, just check
-  });
-
-  return videoExists;
-};
-
-// Function to save generated video as a permanent chat message with attachment
-const saveVideoToChat = async (
-  chatId: string,
-  videoUrl: string,
-  prompt: string,
-  setMessages?: UseChatHelpers['setMessages'],
-  thumbnailUrl?: string,
-) => {
-  if (!setMessages || !chatId) {
-    console.log(
-      '💾 ⚠️ Cannot save video to chat - missing setMessages or chatId',
-    );
-    return;
-  }
-
-  console.log('💾 Saving generated video to chat history...', {
-    chatId,
-    videoUrl: `${videoUrl.substring(0, 50)}...`,
-    prompt,
-  });
-
-  try {
-    // AICODE-NOTE: Check if video already exists in chat to prevent duplicates from clicks
-    const videoExists = checkVideoExistsInChat(setMessages, videoUrl);
-    if (videoExists) {
-      console.log('💾 ⏭️ Video already exists in chat, skipping duplicate save');
-      return;
-    }
-
-    // Create video attachment for permanent storage in chat
-    const videoAttachment = {
-      name: prompt.length > 50 ? `${prompt.substring(0, 50)}...` : prompt, // Use prompt as name
-      url: videoUrl,
-      contentType: 'video/mp4',
-      thumbnailUrl: thumbnailUrl, // Add thumbnail for preview
-    };
-
-    // Create message with video attachment and valid UUID
-    const videoMessage = {
-      id: generateUUID(), // Use proper UUID instead of random string
-      role: 'assistant' as const,
-      content: `Generated video: "${prompt}"`,
-      parts: [
-        {
-          type: 'text' as const,
-          text: `Generated video: "${prompt}"`,
-        },
-      ],
-      experimental_attachments: [videoAttachment],
-      createdAt: new Date(),
-    };
-
-    // Add message to chat history
-    setMessages((prevMessages) => [...prevMessages, videoMessage]);
-
-    console.log('💾 ✅ Video added to chat history locally!');
-
-    // Save to database
-    try {
-      const response = await fetch('/api/save-message', {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-        },
-        body: JSON.stringify({
-          chatId,
-          message: {
-            id: videoMessage.id,
-            role: videoMessage.role,
-            parts: videoMessage.parts,
-            attachments: videoMessage.experimental_attachments,
-            createdAt: videoMessage.createdAt,
-          },
-        }),
-      });
-
-      if (!response.ok) {
-        const errorText = await response.text().catch(() => 'Unknown error');
-        throw new Error(
-          `Failed to save message: ${response.status} - ${errorText}`,
-        );
-      }
-
-      console.log('💾 ✅ Video saved to database successfully!');
-    } catch (dbError) {
-      console.warn(
-        '💾 ⚠️ Failed to save to database, but video is in chat locally:',
-        dbError,
-      );
-      // Don't throw - the video is already in chat locally
-    }
-
-    console.log(
-      '💾 🎬 Video will remain accessible even after closing the artifact',
-    );
-  } catch (error) {
-    console.error('💾 ❌ Failed to save video to chat:', error);
-  }
-};
 
 export function useVideoEffects({
   videoUrl,
@@ -231,7 +113,7 @@ export function useVideoEffects({
 
       // Small delay to ensure artifact is updated first
       setTimeout(() => {
-        saveVideoToChat(chatId, videoUrl, prompt, setMessages);
+        saveMediaToChat(chatId, videoUrl, prompt, setMessages, "video");
       }, 100);
     }
   }, [videoUrl, status, hasInitialized, chatId, setMessages, prompt]);
